fix(management-api): tolerate bigint version columns in migration ledger
项目库中 supabase_migrations.schema_migrations.version 为 bigint（Supabase CLI
约定）时，/database/migrations 全流程失败（operator does not exist:
bigint = text），所有项目迁移被阻断。

- findExistingMigration 改为 version::text 比较，兼容 text/bigint 两种列类型
- record_schema_migration 插入时对版本值做 ::bigint 转换（bigint 列直接匹配，
  text 列经赋值转换兼容）

相关：seagoo-ai 迁移 20260806120000 因此故障无法应用（noop 迁移同样失败）

+++ packages/management-api/src/services/project-migration-role.ts
@@ -64,11 +64,13 @@ export function renderProjectMigrationRoleSql(dbName: string, dbUser: string): s
       INSERT INTO supabase_migrations.schema_migrations
         (version, statements, name, checksum, inserted_at)
       VALUES
-        (migration_version, migration_statements, migration_name, migration_checksum, applied_at);
+        -- 线上台账表的 version 可能是 bigint（Supabase CLI 约定）或 text，
+        -- ::bigint 在两种列类型下均可写入（bigint→text 为赋值转换）。
+        (migration_version::bigint, migration_statements, migration_name, migration_checksum, applied_at);
       INSERT INTO public.schema_migrations
         (version, statements, name, checksum, inserted_at)
       VALUES
-        (migration_version, migration_statements, migration_name, migration_checksum, applied_at);
+        (migration_version::bigint, migration_statements, migration_name, migration_checksum, applied_at);
     END
     $record_migration$;
     REVOKE ALL ON FUNCTION supabase_migrations.record_schema_migration(TEXT, TEXT[], TEXT, TEXT, TEXT)
